TASK-49227 : manage my wallet password" (#119)
