+++ wallet-services/src/main/resources/locale/addon/Wallet_en.properties
@@ -240,12 +240,15 @@ exoplatform.wallet.label.etherBalance=Ether balance
 exoplatform.wallet.label.ethers=Ethers
 exoplatform.wallet.label.ethereumAddress=Ethereum address
 exoplatform.wallet.label.initializationstatus=Initialization
-exoplatform.wallet.label.newWalletPassword=New wallet password
-exoplatform.wallet.label.currentWalletPassword=Current wallet password
+exoplatform.wallet.label.existingPassword=Existing password
+exoplatform.wallet.label.newWalletPassword=New Password
+exoplatform.wallet.label.newWalletPasswordConfirm=New password confirmation
+exoplatform.wallet.label.currentWalletPassword=Enter existing password
 exoplatform.wallet.label.walletPassword=Wallet password
 exoplatform.wallet.label.createWalletInvitation=You are invited to create your wallet
 exoplatform.wallet.label.exoWalletDescription=eXo Wallet is a cryptocurrency wallet using a digital encryption keys to interact with a blockchain network
-exoplatform.wallet.label.walletPasswordPlaceholder=Enter wallet password
+exoplatform.wallet.label.walletPasswordPlaceholder=Enter new password
+exoplatform.wallet.label.walletPasswordPlaceholderConfirm=confirm new password 
 exoplatform.wallet.label.setWalletPasswordPlaceholder=Enter your wallet password
 exoplatform.wallet.label.confirmWalletPasswordPlaceholder=Confirm your wallet Password
 exoplatform.wallet.label.transactionsOfWallet=Transactions of
@@ -394,6 +397,8 @@ exoplatform.wallet.label.yearShortPeriod=Year {0}
 exoplatform.wallet.message.manageDigitalKey=You can transmit this public wallet address or QR code. This will allow you to receive Meeds.
 exoplatform.wallet.message.settingsDescription= In order to secure your Meeds, a digital key is required, we generate and store it for you.You will need to assign a password to it
 exoplatform.wallet.message.managePasswordDescription=Change of password or request of reinitialization in case of loss.
+exoplatform.wallet.message.managePassword=You'll have to enter your existing password
+exoplatform.wallet.message.rememberMyPassword=Your password will not be requested when validating transactions in the online store
 exoplatform.wallet.message.securityPasswordManagement=Add an extra layer of security by using a dedicated password for your wallet.
 exoplatform.wallet.message.encryptionKeysManagement=Keys are used to secure your wallet transactions. We can keep them safe for you on server, or you can manage them yourself.
 exoplatform.wallet.message.transactionInProgress=A transaction is in progress
