+++ wallet-webapps-common/src/main/webapp/css/main.less
@@ -866,6 +866,22 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
   }
 }
 
+.walletManagePassword {
+  .v-list-item__title{
+    white-space: normal !important;
+  }
+}
+.walletManagePasswordDrawer {
+  .managePasswordLabels{
+    .v-label{
+      text-transform: capitalize;
+    }
+  }
+  .v-input--reverse .v-input__slot {
+    flex-direction: row-reverse;
+    justify-content: flex-end;
+  }
+}
 .walletSetting {
   .v-list-item__subtitle{
     white-space: normal !important;
